requirements.txt
streamlit==1.30.0
langchain==0.1.16
langchain-groq==0.1.3
praw==7.8.1
python-dotenv==1.0.1
tornado==6.3.3